+++ example/a.ts
@@ -1,18 +1,9 @@
 import { Elysia } from '../src'
 
-const sdkApp = new Elysia({ prefix: '/sdk' }).mount(
-	'/problems-domain',
-	(request) => {
-		console.log(request.url)
+async function mountedHandler(request: Request): Promise<Response> {
+	return Response.json(await request.json())
+}
 
-		return Response.json({ path: new URL(request.url).pathname })
-	}
-)
-
-const app = new Elysia().use(sdkApp)
-
-const response = await app
-	.handle(new Request('http://localhost/sdk/problems-domain/problems'))
-	.then((x) => x.text())
-
-console.log(response)
+const serverAotFalse = new Elysia({ aot: false })
+	.mount('/api', async (request) => Response.json(await request.json()))
+	.listen(3002)